+++ dev/experiments/crall/expt_drop6_2023-03-09.sh
@@ -9592,30 +9592,27 @@ EXPERIMENT_NAME=Drop7_scratch_V04
 DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
 TARGET_LR=3e-4
 MAX_STEPS=80000
+BATCH_SIZE=20
+MAX_EPOCHS=743
+ACCUMULATE_GRAD_BATCHES=19
+TRAIN_ITEMS_PER_EPOCH=40960
+TRAIN_BATCHES_PER_EPOCH=$(python -c "print($TRAIN_ITEMS_PER_EPOCH // $BATCH_SIZE)")
+echo "TRAIN_BATCHES_PER_EPOCH = $TRAIN_BATCHES_PER_EPOCH"
 
+python -m geowatch.cli.experimental.recommend_size_adjustments \
+    --MAX_STEPS=$MAX_STEPS \
+    --MAX_EPOCHS=$MAX_EPOCHS \
+    --BATCH_SIZE=$BATCH_SIZE \
+    --ACCUMULATE_GRAD_BATCHES=$ACCUMULATE_GRAD_BATCHES \
+    --TRAIN_BATCHES_PER_EPOCH="$TRAIN_BATCHES_PER_EPOCH" \
+    --TRAIN_ITEMS_PER_EPOCH="$TRAIN_ITEMS_PER_EPOCH"
 
 PREV_CHECKPOINT=$(python -m geowatch.cli.experimental.find_recent_checkpoint --default_root_dir="$DEFAULT_ROOT_DIR")
-
-# Find the most recent checkpoint (TODO add utility for this)
-PREV_CHECKPOINT=$(python -c "if 1:
-    import ubelt as ub
-    root_dir = ub.Path('$DEFAULT_ROOT_DIR')
-    checkpoints = list((root_dir / 'lightning_logs').glob('version_*/checkpoints/*.ckpt'))
-    if len(checkpoints) == 0:
-        print('None')
-    else:
-        version_to_checkpoints = ub.group_items(checkpoints, key=lambda x: int(x.parent.parent.name.split('_')[-1]))
-        max_version = max(version_to_checkpoints)
-        candidates = version_to_checkpoints[max_version]
-        checkpoints = sorted(candidates, key=lambda p: p.stat().st_mtime)
-        chosen = checkpoints[-1]
-        print(chosen)
-")
 echo "PREV_CHECKPOINT = $PREV_CHECKPOINT"
 
 WATCH_GRID_WORKERS=4 python -m geowatch.tasks.fusion fit --config "
 data:
-  batch_size              : 20
+  batch_size              : $BATCH_SIZE
   num_workers             : 4
   train_dataset           : $TRAIN_FPATH
   vali_dataset            : $VALI_FPATH
@@ -9639,7 +9636,7 @@ data:
   use_centered_positives  : True
   use_grid_positives      : true
   verbose                 : 1
-  max_epoch_length        : 40960
+  max_epoch_length        : $TRAIN_ITEMS_PER_EPOCH
   mask_low_quality        : false
   mask_samecolor_method   : null
 model:
@@ -9699,7 +9696,7 @@ optimizer:
       - 0.9
       - 0.99
 trainer:
-  accumulate_grad_batches: 19
+  accumulate_grad_batches: $ACCUMULATE_GRAD_BATCHES
   default_root_dir     : $DEFAULT_ROOT_DIR
   accelerator          : gpu
   devices              : 0,
@@ -9713,7 +9710,7 @@ trainer:
   max_epochs: 743
   num_sanity_val_steps: 0
   limit_val_batches: 4096
-  limit_train_batches: 40960
+  limit_train_batches: $TRAIN_BATCHES_PER_EPOCH
   callbacks:
         - class_path: pytorch_lightning.callbacks.ModelCheckpoint
           init_args:
@@ -9724,45 +9721,169 @@ trainer:
               save_last: true
 " --ckpt_path="$PREV_CHECKPOINT"
 
+export DVC_DATA_DPATH=$(geowatch_dvc --tags="phase2_data")
+export DVC_EXPT_DPATH=$(geowatch_dvc --tags="phase2_expt")
+python -m geowatch.mlops.manager "list" --dataset_codes "Drop7-MedianNoWinter10GSD-V2"
 
-python -c "if 1:
-    import sympy
-    limit_train_batches, batch_size, accumulate_grad_batches, max_epochs, MAX_STEPS = sympy.symbols(
-        'limit_train_batches, batch_size, accumulate_grad_batches, max_epochs, MAX_STEPS')
+#python -m geowatch.mlops.manager "add packages" --dataset_codes "Drop7-MedianNoWinter10GSD-V2"
+python -m geowatch.mlops.manager "push packages" --dataset_codes "Drop7-MedianNoWinter10GSD-V2"
+python -m geowatch.mlops.manager "list packages" --dataset_codes "Drop7-MedianNoWinter10GSD-V2"
 
-    import ubelt as ub
 
-    subs = {
-        limit_train_batches: 40960,
-        batch_size: 20,
-        accumulate_grad_batches: 19,
-        max_epochs: 743,
-        MAX_STEPS: 80000,
-    }
+#### ----
+# Yardrat training from scratch (rework params)
 
-    effective_batch_size = accumulate_grad_batches * batch_size
-    #steps_per_epoch = sympy.floor(limit_train_batches / effective_batch_size)
-    steps_per_epoch = limit_train_batches / effective_batch_size
-    total_steps = max_epochs * steps_per_epoch
-    total_steps.subs(subs)
+export CUDA_VISIBLE_DEVICES=0
+DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware='ssd')
+DVC_EXPT_DPATH=$(geowatch_dvc --tags='phase2_expt' --hardware='hdd')
+echo "DVC_EXPT_DPATH = $DVC_EXPT_DPATH"
+WORKDIR=$DVC_EXPT_DPATH/training/$HOSTNAME/$USER
+DATASET_CODE=Drop7-MedianNoWinter10GSD-V2
+KWCOCO_BUNDLE_DPATH=$DVC_DATA_DPATH/$DATASET_CODE
+TRAIN_FPATH=$KWCOCO_BUNDLE_DPATH/data_train_rawbands_split6_hack.kwcoco.zip
+VALI_FPATH=$KWCOCO_BUNDLE_DPATH/data_vali_rawbands_split6.kwcoco.zip
+CHANNELS="(L8,S2,WV,PD):(blue|green|red)"
+EXPERIMENT_NAME=Drop7_scratch_V05
+DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
+TARGET_LR=3e-3
+MAX_STEPS=80000
+BATCH_SIZE=10
+MAX_EPOCHS=743
+ACCUMULATE_GRAD_BATCHES=38
+TRAIN_ITEMS_PER_EPOCH=40960
+TRAIN_BATCHES_PER_EPOCH=$(python -c "print($TRAIN_ITEMS_PER_EPOCH // $BATCH_SIZE)")
+echo "TRAIN_BATCHES_PER_EPOCH = $TRAIN_BATCHES_PER_EPOCH"
 
-    # The training progress iterator should show this number as the total number
-    train_epoch_prog_iters = sympy.ceiling((limit_train_batches / batch_size).subs(subs).evalf())
+python -m geowatch.cli.experimental.recommend_size_adjustments \
+    --MAX_STEPS=$MAX_STEPS \
+    --MAX_EPOCHS=$MAX_EPOCHS \
+    --BATCH_SIZE=$BATCH_SIZE \
+    --ACCUMULATE_GRAD_BATCHES=$ACCUMULATE_GRAD_BATCHES \
+    --TRAIN_BATCHES_PER_EPOCH="$TRAIN_BATCHES_PER_EPOCH" \
+    --TRAIN_ITEMS_PER_EPOCH="$TRAIN_ITEMS_PER_EPOCH"
 
-    diff = MAX_STEPS - total_steps
-    curr_diff = diff.subs(subs)
-    print(f'curr_diff={curr_diff.evalf()}')
+PREV_CHECKPOINT=$(python -m geowatch.cli.experimental.find_recent_checkpoint --default_root_dir="$DEFAULT_ROOT_DIR")
+echo "PREV_CHECKPOINT = $PREV_CHECKPOINT"
 
-    if curr_diff > 0:
-        print('Not enough total steps to fill MAX_STEPS')
+# shellcheck disable=SC2037
+CKPT_ARG=python -c "if 1:
+    if '$PREV_CHECKPOINT' == 'None':
+        return ''
     else:
-        print('MAX STEPS will stop training short')
-
-    for k, v in subs.items():
-        print('--- Possible Adjustment For ---')
-        print(k)
-        tmp_subs = (ub.udict(subs) - {k})
-        solutions = sympy.solve(diff.subs(tmp_subs), k)
-        solutions = [s.evalf() for s in solutions]
-        print(solutions)
+        return '--ckpt_fpath=$PREV_CHECKPOINT'
 "
+echo "CKPT_ARG = $CKPT_ARG"
+
+# shellcheck disable=SC2086
+WATCH_GRID_WORKERS=4 python -m geowatch.tasks.fusion fit --config "
+data:
+  batch_size              : $BATCH_SIZE
+  num_workers             : 8
+  train_dataset           : $TRAIN_FPATH
+  vali_dataset            : $VALI_FPATH
+  time_steps              : 9
+  chip_dims               : 256,256
+  window_space_scale      : 10.0GSD
+  input_space_scale       : 10.0GSD
+  output_space_scale      : 10.0GSD
+  channels                : '$CHANNELS'
+  chip_overlap            : 0
+  dist_weights            : 0
+  min_spacetime_weight    : 0.6
+  neg_to_pos_ratio        : 1.0
+  normalize_inputs        : 1024
+  normalize_perframe      : false
+  resample_invalid_frames : 3
+  temporal_dropout        : 0.5
+  time_sampling           : uniform-soft5-soft4-contiguous
+  time_kernel             : '(-3y,-2.5y,-2y,-1.5y,-1y,0,1y,1.5y,2y,2.5y,3y)'
+  upweight_centers        : true
+  use_centered_positives  : True
+  use_grid_positives      : true
+  verbose                 : 1
+  max_epoch_length        : $TRAIN_ITEMS_PER_EPOCH
+  mask_low_quality        : false
+  mask_samecolor_method   : null
+model:
+  class_path: watch.tasks.fusion.methods.MultimodalTransformer
+  init_args:
+    arch_name: smt_it_stm_p16
+    attention_impl: exact
+    attention_kwargs: null
+    backbone_depth: null
+    change_head_hidden: 6
+    change_loss: cce
+    class_head_hidden: 6
+    class_loss: dicefocal
+    class_weights: auto
+    config: null
+    continual_learning: true
+    decoder: mlp
+    decouple_resolution: false
+    dropout: 0.1
+    focal_gamma: 2.0
+    global_change_weight: 0.0
+    global_class_weight: 0.000
+    global_saliency_weight: 1.00
+    input_channels: null
+    input_sensorchan: null
+    learning_rate: 0.001
+    lr_scheduler: CosineAnnealingLR
+    modulate_class_weights: ''
+    multimodal_reduce: learned_linear
+    name: unnamed_model
+    negative_change_weight: 0.01
+    ohem_ratio: null
+    optimizer: RAdam
+    perterb_scale: 1.0e-07
+    positional_dims: 48
+    positive_change_weight: 1
+    rescale_nans: null
+    saliency_head_hidden: 6
+    saliency_loss: focal
+    saliency_weights: auto
+    stream_channels: 16
+    tokenizer: linconv
+
+lr_scheduler:
+  class_path: torch.optim.lr_scheduler.OneCycleLR
+  init_args:
+    max_lr: $TARGET_LR
+    total_steps: $MAX_STEPS
+    anneal_strategy: cos
+    pct_start: 0.3
+optimizer:
+  class_path: torch.optim.AdamW
+  init_args:
+    lr: $TARGET_LR
+    weight_decay: 3e-06
+    betas:
+      - 0.9
+      - 0.99
+trainer:
+  accumulate_grad_batches: $ACCUMULATE_GRAD_BATCHES
+  default_root_dir     : $DEFAULT_ROOT_DIR
+  accelerator          : gpu
+  devices              : 0,
+  #devices              : 0,1
+  #strategy             : ddp
+  check_val_every_n_epoch: 1
+  enable_checkpointing: true
+  enable_model_summary: true
+  log_every_n_steps: 1
+  logger: true
+  max_epochs: 743
+  num_sanity_val_steps: 0
+  limit_val_batches: 4096
+  limit_train_batches: $TRAIN_BATCHES_PER_EPOCH
+  callbacks:
+        - class_path: pytorch_lightning.callbacks.ModelCheckpoint
+          init_args:
+              monitor: val_loss
+              mode: min
+              save_top_k: 5
+              filename: '{epoch:04d}-{step:06d}-{val_loss:.3f}.ckpt'
+              save_last: true
+initializer:
+    init: noop
+" $CKPT_ARG