+++ scripts/prepare_drop8_L2.sh
@@ -20,7 +20,8 @@ DVC_DATA_DPATH=$(geowatch_dvc --tags=phase3_data --hardware="hdd")
 
 
 #SENSORS="ta1-ls-ara-4,ta1-pd-ara-4,ta1-s2-ara-4,ta1-wv-ara-4"
-SENSORS="sentinel-s2-l2a-cogs,landsat-c2l2-sr,planet-dove,worldview-nitf"
+#SENSORS="sentinel-s2-l2a-cogs,landsat-c2l2-sr,planet-dove,worldview-nitf"
+SENSORS="sentinel-s2-l2a-cogs,landsat-c2l2-sr,planet-dove"
 #SENSORS="sentinel-s2-l2a-cogs,landsat-c2l2-sr,planet-dove,worldview-nitf"
 
 DATASET_SUFFIX=Drop8-L2
@@ -30,8 +31,8 @@ DATASET_SUFFIX=Drop8-L2
 #SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop6_hard_v1/site_models/KR_R001*.geojson"
 
 # All Regions
-REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8/region_models/*.geojson"
-SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8/site_models/*.geojson"
+REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8/region_models/*_*0*.geojson"
+SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8/site_models/*_*0*_*.geojson"
 
 # T&E Regions Only
 #REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8/region_models/*_R*.geojson"
@@ -67,7 +68,7 @@ python -m geowatch.cli.prepare_ta2_dataset \
     --target_gsd="10GSD" \
     --cache=0 \
     --verbose=100 \
-    --skip_existing=0 \
+    --skip_existing=1 \
     --force_min_gsd=2.0 \
     --force_nodata=-9999 \
     --align_tries=1 \